Reorganizar directorio hooks
